web/requirements.txt
WTForms==2.3.3
pygithub3==0.5.1
itsdangerous==1.1.0
onetimepass==1.0.1
PyQRCode==1.2.1
SQLAlchemy==1.3.23
flask_expects_json==1.5.0
Flask_SocketIO==5.0.1
Flask_Session==0.3.2
Flask==1.1.2
Flask_SQLAlchemy==2.4.4
Jinja2==2.11.3
Flask_WTF==0.14.3
requests==2.25.1
rsa==4.7.2
Flask_Login==0.5.0
Werkzeug==1.0.1
Pillow==8.2.0
PyYAML==5.4.1
truffleHog==2.2.1
